+++ lib/src/qr_scanner_overlay_shape.dart
